+++ package/kedro_viz/api/__init__.py
@@ -1,1 +1,1 @@
-"""`kedro_viz.api` defines the FastAPI app to serve Kedro data in a RESTful API."""
+"""`kedro_viz.api` defines the FastAPI app to serve Kedro data."""
