+++ package/kedro_viz/api/graphql/__init__.py
@@ -1,0 +1,2 @@
+# TODO: docstrings
+from .router import router
