+++ package/kedro_viz/api/rest/__init__.py
@@ -1,0 +1,1 @@
+from .router import router
